requirements.txt
# YAML:
pyyaml
#dict2xml
# Timezones:
pytz
python-dateutil
# Command line:
click
# VS Code:
pylint
black
